青木修正

+++ tools/master/importStationKind.py
@@ -12,13 +12,17 @@ class importStationKind(webapp2.RequestHandler):
     #FILE_PATH = "/miral-1265.appspot.com/initdata/station20160401free.csv"
     FILE_PATH = "testdata/station20160401free.csv"
     
-    def outDB(self):
+    def outDB(self, delflg):
         log = MiralLogger()
         
         #データクリア
-        ndb.delete_multi(
-            MstStationKind.query().fetch(keys_only=True)
-        )    
+        if delflg:
+            log.debug("delete!!!!!!!!!!")
+            ndb.delete_multi(
+                MstStationKind.query().fetch(keys_only=True)
+            )    
+            
+            return 0
             
         #with gcs.open(self.FILE_PATH, 'r') as f:
         with open(self.FILE_PATH, 'r') as f:
@@ -26,16 +30,18 @@ class importStationKind(webapp2.RequestHandler):
             
             next(reader)
             
-            #for c in range(MstStationKind().query().count()):
-            #    next(reader)
+            for c in range(MstStationKind().query().count()):
+                next(reader)
                 
             #ターゲット路線
             targetTrainList=[]
+            targetTrainKeyList={}
             
             qry = MstTrainKind.query()
             
             for tr in qry.fetch():
                 targetTrainList.append(str(tr.trainCd))
+                targetTrainKeyList[str(tr.trainCd)]=tr.key
             
             c2=0
             
@@ -50,16 +56,16 @@ class importStationKind(webapp2.RequestHandler):
                 station.stationCd=int(row[0])
                 station.stationGpCd=int(row[1])
                 station.stationName=row[2]
-                station.trainCd=int(row[5])
+                station.trainKey=targetTrainKeyList[row[5]]
                 station.geoCd=ndb.GeoPt(float(row[10]),float(row[9]))
                 station.displayOrder=int(row[1])
 
                 station.put()
                 
                 c2=c2+1
-                
-                #if c2 > 1000:
-                #    break
+                 
+                if c2 > 1000:
+                    break
             
             return c2
         
@@ -68,9 +74,12 @@ class importStationKind(webapp2.RequestHandler):
         
     def get(self):
         
-        #p = self.request.GET['s']
+        d = self.request.GET['delete']
         
-        c = self.outDB()
-        
+        if d=="1":
+            c = self.outDB(d)
+        else:
+            c = self.outDB(False)
+            
         self.response.out.write(u'終了! ' + str(c) + " "+str(MstStationKind().query().count()))
